Restructure component files. TitleInput, Save / Delete button components live in Header since that component uses them.
