+++ src/pages/Home/components/RiffinEditor/components/Header/components/DeleteTabButton/DeleteTabButton.js
